Initial draft of the seach feature
