+++ app/lib/util/library-navigator/index.dart
@@ -31,7 +31,7 @@ class LibraryNavigator extends RouterDelegate
         pages: [
           MaterialPage(
               key: ValueKey("PrimarySectionsRoute"),
-              child: PrimarySectionsRoute()),
+              child: PrimarySectionRoute()),
           for (final book
               in appState.sections.where((section) => section.wasNavigatedTo))
             MaterialPage(
