+++ inside_data/bin/1.1allWpJson.dart
@@ -1,7 +1,7 @@
 import 'dart:convert';
 
 import 'package:inside_data/inside_data.dart';
-import 'package:inside_data/src/loaders/wordpress/wordpress_repository.dart';
+import 'package:inside_data/src/wordpress/wordpress_repository.dart';
 
 Future<void> main(List<String> args) async {
   const domain = 'insidechassidus.org';
