+++ inside_data/lib/inside_data.dart
@@ -2,4 +2,5 @@ library inside_data;
 
 export 'package:inside_data/src/inside_data.dart';
 export 'package:inside_data/src/loaders/wordpress_loader.dart';
+export 'package:inside_data/src/wordpress/wordpress_search.dart';
 export 'package:inside_data/src/constants.dart';
